fix: Formatted code.

+++ script/MyNFT.s.sol
@@ -8,14 +8,10 @@ contract MyNFTScript is Script {
     function setUp() public {}
 
     function run() public {
-        uint privateKey = uint256(vm.envBytes32("TEST_ACCOUNT_1_PRIVATE_KEY"));
+        uint256 privateKey = uint256(vm.envBytes32("TEST_ACCOUNT_1_PRIVATE_KEY"));
         vm.startBroadcast(privateKey);
 
-        MyNFT mynft = new MyNFT(
-            "BestNft",
-            "---@@66@@---",
-            "https://example.com/token-image.png"
-        );
+        MyNFT mynft = new MyNFT("BestNft", "---@@66@@---", "https://example.com/token-image.png");
 
         mynft.togglePublicSale();
